Move cleanup related test utils and tests
