+++ src/storage/mvcc/mod.rs
@@ -320,7 +320,6 @@ pub fn default_not_found_error(key: Vec<u8>, hint: &str) -> Error {
 pub mod tests {
     use super::*;
     use crate::storage::kv::{Engine, Modify, ScanMode, Snapshot, WriteData};
-    use crate::storage::txn::cleanup;
     use concurrency_manager::ConcurrencyManager;
     use engine_traits::CF_WRITE;
     use kvproto::kvrpcpb::{Context, IsolationLevel};
@@ -409,35 +408,6 @@ pub mod tests {
         let cm = ConcurrencyManager::new(start_ts);
         let mut txn = MvccTxn::new(snapshot, start_ts, true, cm);
         assert!(txn.rollback(Key::from_raw(key)).is_err());
-    }
-
-    pub fn must_cleanup<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        start_ts: impl Into<TimeStamp>,
-        current_ts: impl Into<TimeStamp>,
-    ) {
-        let ctx = Context::default();
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let current_ts = current_ts.into();
-        let cm = ConcurrencyManager::new(current_ts);
-        let mut txn = MvccTxn::new(snapshot, start_ts.into(), true, cm);
-        cleanup(&mut txn, Key::from_raw(key), current_ts, true).unwrap();
-        write(engine, &ctx, txn.into_modifies());
-    }
-
-    pub fn must_cleanup_err<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        start_ts: impl Into<TimeStamp>,
-        current_ts: impl Into<TimeStamp>,
-    ) -> Error {
-        let ctx = Context::default();
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let current_ts = current_ts.into();
-        let cm = ConcurrencyManager::new(current_ts);
-        let mut txn = MvccTxn::new(snapshot, start_ts.into(), true, cm);
-        cleanup(&mut txn, Key::from_raw(key), current_ts, true).unwrap_err()
     }
 
     pub fn must_gc<E: Engine>(engine: &E, key: &[u8], safe_point: impl Into<TimeStamp>) {
