+++ src/storage/txn/mod.rs
@@ -245,6 +245,7 @@ pub mod tests {
         must_succeed_return_value as must_acquire_pessimistic_lock_return_value,
         must_succeed_with_ttl as must_acquire_pessimistic_lock_with_ttl,
     };
+    pub use actions::cleanup::tests::{must_err as must_cleanup_err, must_succeed as must_cleanup};
     pub use actions::commit::tests::{must_err as must_commit_err, must_succeed as must_commit};
     pub use actions::pessimistic_prewrite::tests::try_pessimistic_prewrite_check_not_exists;
     pub use actions::prewrite::tests::{try_prewrite_check_not_exists, try_prewrite_insert};
